Split up components in to individual files
