+++ src/event.rs
@@ -1,0 +1,69 @@
+use chrono::NaiveDateTime;
+use serde::Serialize;
+use std::collections::HashMap;
+
+use crate::error::Error;
+
+// This exists so that the client doesn't have to specify the API key over and over
+#[derive(Serialize)]
+pub(crate) struct InnerEvent {
+    api_key: String,
+    event: String,
+    properties: Properties,
+    timestamp: Option<NaiveDateTime>,
+}
+
+impl InnerEvent {
+    pub(crate) fn new(event: Event, api_key: String) -> Self {
+        Self {
+            api_key,
+            event: event.event,
+            properties: event.properties,
+            timestamp: event.timestamp,
+        }
+    }
+}
+
+#[derive(Serialize, Debug, PartialEq, Eq)]
+pub struct Event {
+    event: String,
+    properties: Properties,
+    timestamp: Option<NaiveDateTime>,
+}
+
+#[derive(Serialize, Debug, PartialEq, Eq)]
+pub struct Properties {
+    distinct_id: String,
+    props: HashMap<String, serde_json::Value>,
+}
+
+impl Properties {
+    fn new<S: Into<String>>(distinct_id: S) -> Self {
+        Self {
+            distinct_id: distinct_id.into(),
+            props: Default::default(),
+        }
+    }
+}
+
+impl Event {
+    pub fn new<S: Into<String>>(event: S, distinct_id: S) -> Self {
+        Self {
+            event: event.into(),
+            properties: Properties::new(distinct_id),
+            timestamp: None,
+        }
+    }
+
+    /// Errors if `prop` fails to serialize
+    pub fn insert_prop<K: Into<String>, P: Serialize>(
+        &mut self,
+        key: K,
+        prop: P,
+    ) -> Result<(), Error> {
+        let as_json =
+            serde_json::to_value(prop).map_err(|e| Error::Serialization(e.to_string()))?;
+        let _ = self.properties.props.insert(key.into(), as_json);
+        Ok(())
+    }
+}
